+++ src/auth/signin.ts
@@ -34,7 +34,7 @@ async function fetchSignIn(
     passwordHashSalt,
     32,
     ARGON2ID_ITERATIONS,
-    32768 * 1024,
+    BigInt(32768 * 1024),
   );
 
   const authPasswordTokenBuffer = new HKDF(
