+++ src/auth/signup.ts
@@ -21,7 +21,7 @@ export default async function signUp(username: string, password: string): Promis
     profileData.passwordHashSalt,
     32,
     ARGON2ID_ITERATIONS,
-    32768 * 1024,
+    BigInt(32768 * 1024),
   );
 
   const authPasswordTokenBuffer = new HKDF(
